worker/pyproject.toml
[project]
name = "ai-news-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "anthropic>=0.49",
    "openai>=1.70",
    "httpx>=0.28",
    "feedparser>=6.0",
    "supabase>=2.0",
    "celery[redis]>=5.3",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "bleach>=6.0",
    "langsmith>=0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.12"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
